Simple PopUp
Start with popup that user can get rid of by clicking the x at the corner. Still have some sizing issues and still need to consider how to turn the player's movement off when in the pop up window. Goal is to have player start in the cell and be able to click on things in the cell that allow it to exit the cell and then enter the mode where they can move around.