Button in PopUp Prints to Screen
Still want it to do something substantial. I think we also need to be thinking about how the scenes will fit together. This pop I have with a picture probably won't be the intended use. Not completely sure how we'll use them yet. Maybe in the cafe when going up to players.