Window now stops player input when open and continues player movement when closed
Signals are very important here. There a signal connect the test pop up and the player so that controls if the player input is picked up or not. Most signals need to be emitted, but for some reason some of the built in signals for windows like (about to pop up or closed requested ) don't need to be emitted. I think this is because source and target is the popup window. Had to add an extra popup() call that seemed to make the signal between the pop up and player scene work.