Door Access Completed
Going to give up on getting the full door_passcode scene text change for now. I don't understand why it's either text changes after entering and then nothing or no text changes but it goes to the next scene. I want a text change and then it move to the next scene. :(